requirements/runtime.txt
# Add requirements here

ubelt >= 1.3.0
scriptconfig>=0.7.8  ;                            python_version >= '3.6'    # Python 3.6+
rich>=12.4.4         ;                            python_version >= '3.8'    # Python 3.8+
rich>=11.2.0         ; python_version < '3.8' and python_version >= '3.6'    # Python 3.6 
git-python>=1.0.0    ;                            python_version >= '2.7'    # Python 2.7+
packaging>=21.3

networkx>=2.8           ; python_version < '4.0'   and python_version >= '3.11'   # Python 3.11+
networkx>=2.7           ; python_version < '3.11'  and python_version >= '3.8'    # Python 3.8-3.11
networkx>=2.6.2         ; python_version < '3.8'   and python_version >= '3.7'    # Python 3.7
networkx>=2.2.0,<=2.5.1 ; python_version < '3.7.0' and python_version >= '3.6.0'  # Python 3.6
